fix(gui): change EnumValueViewModel.NumericValue from uint to long
Replace uint with long for NumericValue property to support all enum underlying types including signed integers (int, sbyte, short) and larger unsigned values (uint with values > int.MaxValue). This fixes handling of enums like BuildingTabType with negative values (-1) and ConnectorType with large uint values (0x80000000).

+++ EarthTool.PAR.GUI/ViewModels/EnumPropertyEditorViewModel.cs
@@ -156,7 +156,7 @@ public class EnumPropertyEditorViewModel : PropertyEditorViewModel
       {
         Value = value,
         DisplayName = FormatEnumName(name),
-        NumericValue = Convert.ToUInt32(value),
+        NumericValue = Convert.ToInt64(value),
         Description = name // TODO: Get from Description attribute if available
       };
 
@@ -193,9 +193,9 @@ public class EnumValueViewModel
   public string DisplayName { get; set; } = string.Empty;
 
   /// <summary>
-  /// Gets or sets the numeric value.
+  /// Gets or sets the numeric value (supports all enum underlying types).
   /// </summary>
-  public uint NumericValue { get; set; }
+  public long NumericValue { get; set; }
 
   /// <summary>
   /// Gets or sets the description.
